Fix two length underflows that cause an unbounded buffer over-read
Two places compute len - 1 on a length a document can drive to zero and
pass the result to rb_intern3() as a long. The subtraction underflows
and the length becomes enormous while the pointer still points into the
document, so each one is an unbounded buffer over-read that walks
forward until it leaves mapped memory. Both crash the process.

The value of a ^m tag is a colon followed by the symbol name, so
hat_cstr() skips the colon with rb_intern3(str + 1, len - 1, ...). An
empty value leaves nothing to skip: len is a size_t, so len - 1 is
SIZE_MAX and the symbol lookup hashes from one byte past the value until
it reaches an unmapped page. str_to_value() guards the same expression
with 0 < len; hat_cstr() did not. This needs no options at all, since
object mode is the shipped default, so any application calling Oj.load
on a document it did not write could be crashed by nine bytes.

Val.klen was a uint16_t while read_str() assigns pi->cur - str, so a key
of exactly 65,536 bytes wrapped to zero while Val.k1 still held its first
character. When that character is a colon, calc_hash_key() takes the
symbol branch and evaluates rb_intern3(kval->key + 1, kval->klen - 1),
where the subtraction promotes to -1. The same truncation silently
mangled long keys in every mode that goes through read_str(): JSON puts
no limit on the length of a key, but a 65,536 byte one parsed as the
empty string.

Val.klen and Val.clen become size_t, so the length of a key is whatever
the document says it is and there is nothing left to wrap. The locals
that copied them into an int follow, along with two casts that only
existed to compare an int with the old narrow field. struct _val grows
from 88 to 104 bytes, so a ParseInfo on the C stack goes from 5,664 to
6,688 bytes.

Both subtractions are guarded as well, since a guard is cheaper than
proving no other caller reaches them with zero. hat_cstr() returns 0 for
an empty ^m value, which is what it already does for a tag it cannot
handle, so {"^m":""} parses as the plain hash {"^m" => ""}.

Rejecting keys that do not fit, the way Oj::Parser does for its own
int16_t length, was the alternative. It keeps the struct small, but
65,535 is not a limit JSON has and the field would be the only reason
for it.

{"^m":":abc"} still parses as :abc. Keys of any length now parse as what
the document says rather than truncated, and nothing that parsed
correctly before parses differently now.

Both tests crash the interpreter without the fix rather than failing, so
they guard against a regression instead of asserting something that
passes either way.

+++ ext/oj/val_stack.h
@@ -35,8 +35,8 @@ typedef struct _val {
     const char    *classname;
     VALUE          clas;
     OddArgs        odd_args;
-    uint16_t       klen;
-    uint16_t       clen;
+    size_t         klen;
+    size_t         clen;
     char           next;  // ValNext
     char           k1;    // first original character in the key
     char           kalloc;
